Cryptography/Kinda_AESthetic/Docker/app/requirements.txt
pycryptodome==3.12.0
